Extract useList hook
